Move Layer and Source, and start default styling
