+++ packages/viz/src/declarations/declarations.d.ts
@@ -1,2 +1,9 @@
-// TODO: Add proper module declaration
-declare module '@deck.gl/geo-layers';
+// TODO: Add proper module declaration;
+import { TileLayer } from '@deck.gl/geo-layers';
+import { GeoJsonLayerProps } from '@deck.gl/layers/geojson-layer/geojson-layer';
+
+declare module '@deck.gl/geo-layers' {
+  export class MVTTileLayer<D> extends TileLayer<D> {
+    constructor(props: GeoJsonLayerProps<D>);
+  }
+}
