+++ packages/viz/src/lib/Layer.ts
@@ -3,10 +3,14 @@ import { MapOptions, Maps } from '@carto/toolkit-maps';
 import { MVTTileLayer } from '@deck.gl/geo-layers';
 
 import Source from './Source';
+import Styles from './style/Style';
 
 const defaultMapOptions: MapOptions = {
   vector_extent: 2048,
-  vector_simplify_extent: 2048
+  vector_simplify_extent: 2048,
+  metadata: {
+    geometryType: true
+  }
 };
 
 export class Layer {
@@ -14,16 +18,20 @@ export class Layer {
   private _credentials: Credentials;
 
   private _layerSource: Source;
+  private _layerStyles: Styles;
   private _layerOptions: MapOptions;
   private _layerInstantiation: Promise<any>; // TODO: Change to a proper definition
 
-  constructor(source: string, options: LayerOptions = {}) {
+  constructor(source: string, styles = {}, options: LayerOptions = {}) {
+    // TODO: Add property validation
     const { mapOptions, credentials = defaultCredentials } = options;
 
     this._credentials = credentials;
     this._mapsClientInstance = new Maps(this._credentials);
 
     this._layerSource = new Source(source);
+    this._layerStyles = new Styles(styles);
+
     this._layerOptions = Object.assign({}, defaultMapOptions, mapOptions);
     this._layerInstantiation = this._mapsClientInstance.instantiateMapFrom(
       buildInstantiationOptions({ mapOptions: this._layerOptions, mapSource: this._layerSource })
@@ -31,8 +39,8 @@ export class Layer {
   }
 
   public async addTo(deckInstance: any) {
-    const createdDeckLayer = await this.getDeckGLLayer();
     const currentDeckLayers = deckInstance.props.layers;
+    const createdDeckLayer = await this.getDeckGLLayer();
 
     deckInstance.setProps({
       layers: [
@@ -42,10 +50,9 @@ export class Layer {
     });
   }
 
-  public async getDeckGLLayer(layerProps: { layerType?: any } = {}) {
+  public async getDeckGLLayer() {
     // TODO: Parse through Babel
-    const { layerType: sublayerType, ...overridenStyleProps} = layerProps;
-    const { urlTemplates } = await this._layerInstantiation.then(this._parseInstantiationResult);
+    const { urlTemplates, geometryType } = await this._layerInstantiation.then(this._parseInstantiationResult);
 
     const defaultLayerProps = {
       getLineColor: [192, 0, 0],
@@ -56,18 +63,25 @@ export class Layer {
       uniquePropertyName: 'cartodb_id'
     };
 
-    const layerProperties = Object.assign({}, defaultLayerProps, overridenStyleProps);
+    const layerProperties = Object.assign({},
+      defaultLayerProps,
+      this._layerStyles.getProperties(),
+    );
+
     return new MVTTileLayer(layerProperties);
   }
 
   private _parseInstantiationResult(instantiationData: any) {
-    const urlData = instantiationData.metadata.url.vector;
+    const metadata = instantiationData.metadata;
 
+    const urlData = metadata.url.vector;
     const urlTemplates = urlData.subdomains.map(
       (subdomain: string) => urlData.urlTemplate.replace('{s}', subdomain)
     );
 
-    return { urlTemplates };
+    const geometryType = metadata.layers[0].meta.stats.geometryType.split('ST_')[1];
+
+    return { urlTemplates, geometryType };
   }
 
   public get credentials() {
